pyproject.toml
[project]
name = "criando-agentes-com-agno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "agno>=2.3.21",
    "groq>=1.0.0",
    "python-dotenv>=1.2.1",
]
